+++ rust/main/lander/src/adapter/chains.rs
@@ -1,5 +1,6 @@
 pub use ethereum::EthereumTxPrecursor;
 pub use factory::AdapterFactory;
+pub use radix::RadixTxPrecursor;
 pub use sealevel::SealevelTxPrecursor;
 
 mod factory;
@@ -7,4 +8,5 @@ mod factory;
 // chains modules below
 mod cosmos;
 pub mod ethereum;
+pub mod radix;
 pub mod sealevel;