+++ solidity/contracts/upgrade/TransparentUpgradeableProxy.sol
@@ -3,4 +3,4 @@
 
 pragma solidity ^0.8.0;
 
-import "@openzeppelin/contracts/proxy/transparent/TransparentUpgradeableProxy.sol";
+import {TransparentUpgradeableProxy} from "@openzeppelin/contracts/proxy/transparent/TransparentUpgradeableProxy.sol";