+++ typescript/cli/src/config/protocols.ts
@@ -1,0 +1,6 @@
+import { ProtocolType } from '@hyperlane-xyz/utils';
+
+export const COMPATIBLE_PROTOCOLS = [
+  ProtocolType.Ethereum,
+  ProtocolType.CosmosNative,
+];