+++ typescript/sdk/src/index.ts
@@ -121,10 +121,10 @@ export {
   BuildArtifact,
   CompilerOptions,
   ContractVerificationInput,
-  ExplorerLicenseType,
   VerificationInput,
 } from './deploy/verify/types.js';
 export * as verificationUtils from './deploy/verify/utils.js';
+export { ExplorerLicenseType } from './block-explorer/etherscan.js';
 export { ZKSyncContractVerifier } from './deploy/verify/ZKSyncContractVerifier.js';
 export { executeWarpDeploy, enrollCrossChainRouters } from './deploy/warp.js';
 export {
@@ -385,7 +385,7 @@ export {
   ViemProvider,
   ViemTransaction,
   ViemTransactionReceipt,
-  GroupedTransactions,
+  TypedAnnotatedTransaction,
   ProtocolTypedTransaction,
   RadixSDKTransaction,
   RadixTransaction,
@@ -431,6 +431,7 @@ export {
   EV5ImpersonatedAccountTxSubmitterProps,
   EV5ImpersonatedAccountTxSubmitterPropsSchema,
   EvmIcaTxSubmitterProps,
+  isJsonRpcSubmitterConfig,
 } from './providers/transactions/submitter/ethersV5/types.js';
 
 export { TxSubmitterBuilder } from './providers/transactions/submitter/builder/TxSubmitterBuilder.js';
@@ -671,6 +672,7 @@ export {
   LOCKBOX_STANDARDS,
   MINT_LIMITED_STANDARDS,
   PROTOCOL_TO_NATIVE_STANDARD,
+  PROTOCOL_TO_HYP_NATIVE_STANDARD,
   TOKEN_COLLATERALIZED_STANDARDS,
   TOKEN_COSMWASM_STANDARDS,
   TOKEN_HYP_STANDARDS,
@@ -813,3 +815,12 @@ export {
 } from './timelock/evm/constants.js';
 export { EvmEventLogsReader } from './rpc/evm/EvmEventLogsReader.js';
 export { getTimelockExecutableTransactionFromBatch } from './timelock/evm/utils.js';
+export { RadixHookReader } from './hook/RadixHookReader.js';
+export { RadixHookModule } from './hook/RadixHookModule.js';
+export { RadixIsmReader } from './ism/RadixIsmReader.js';
+export { RadixCoreReader } from './core/RadixCoreReader.js';
+export { RadixIsmModule } from './ism/RadixIsmModule.js';
+export {
+  getSignerForChain,
+  MultiProtocolSignerSignerAccountInfo,
+} from './signers/signers.js';